+++ src/Pagination/Concerns/BuildsUrls.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Tobyz\JsonApiServer\Pagination\Concerns;
+
+use Tobyz\JsonApiServer\Context;
+
+trait BuildsUrls
+{
+    private function buildUrl(array $params, Context $context): string
+    {
+        [$selfUrl] = explode('?', $context->request->getUri(), 2);
+
+        $queryParams = array_replace_recursive($context->request->getQueryParams(), $params);
+
+        if (isset($queryParams['filter'])) {
+            foreach ($queryParams['filter'] as &$v) {
+                $v = $v === null ? '' : $v;
+            }
+        }
+
+        ksort($queryParams);
+
+        $queryString = http_build_query($queryParams, '', '&', PHP_QUERY_RFC3986);
+
+        return $selfUrl . ($queryString ? '?' . $queryString : '');
+    }
+}