+++ src/Resource/Paginatable.php
@@ -2,12 +2,13 @@
 
 namespace Tobyz\JsonApiServer\Resource;
 
-use Tobyz\JsonApiServer\Pagination\OffsetPagination;
+use Tobyz\JsonApiServer\Context;
+use Tobyz\JsonApiServer\Pagination\Page;
 
 interface Paginatable
 {
     /**
-     * Paginate the given query.
+     * Get a page of results from the given query.
      */
-    public function paginate(object $query, OffsetPagination $pagination): void;
+    public function paginate(object $query, int $offset, int $limit, Context $context): Page;
 }